+++ controller/authController.js
@@ -1,0 +1,25 @@
+'use strict';
+const jwt = require('express-jwt');
+
+const getTokenFromHeaders = (req) => {
+  const {headers: {authorization}} = req;
+
+  if (authorization && authorization.split(' ')[0] === 'Token') {
+    return authorization.split(' ')[1];
+  }
+  return null;
+};
+
+exports.auth = {
+  required: jwt({
+    secret: 'secret',
+    userProperty: 'payload',
+    getToken: getTokenFromHeaders,
+  }),
+  optional: jwt({
+    secret: 'secret',
+    userProperty: 'payload',
+    getToken: getTokenFromHeaders,
+    credentialsRequired: false,
+  }),
+};
